range_XXp_area moves symmetrically outward Also fix tests, add range_20p_area, add edge case where center sample takes us over threshold.

+++ tests/test_datastructure.py
@@ -130,14 +130,14 @@ class TestDatastructure(unittest.TestCase):
         self.assertEqual(len(w.samples),
                          len(samples))
         self.assertIsInstance(w.samples, np.ndarray)
-        self.assertEqual(w.samples.dtype, np.float64)
+        self.assertEqual(w.samples.dtype, np.float32)
 
-        w.samples = np.array(samples, dtype=np.float64)
+        w.samples = np.array(samples, dtype=np.float32)
 
         self.assertEqual(len(w.samples),
                          len(samples))
         self.assertIsInstance(w.samples, np.ndarray)
-        self.assertEqual(w.samples.dtype, np.float64)
+        self.assertEqual(w.samples.dtype, np.float32)
 
 
 if __name__ == '__main__':
